Allow running Prog::Test::PostgresResource for AWS in development/staging
This adds Config.local_e2e_postgres_test_project_id, which should be
set to a valid project id in the environment if trying this. If set,
instead of creating a new project, the test will create the Postgres
resource in the given project, and it will not destroy the project
after the test completes (it will still destroy the postgres resource
and the created private subnet).

That project should have a us-west-2 location configured with an
appropriate AWS location credential. It will use this location instead
of creating a new location.

This may work with metal providers, but I didn't test that as it
requires additional setup (minio, images, etc.).

+++ prog/test/postgres_base.rb
@@ -22,7 +22,7 @@ class Prog::Test::PostgresBase < Prog::Test::Base
   end
 
   def finish_test(success_msg)
-    postgres_test_project.destroy
+    postgres_test_project.destroy unless Config.local_e2e_postgres_test_project_id
     fail_test(frame["fail_message"]) if frame["fail_message"]
     pop success_msg
   end
